refactor(config): move MongoDB connection logic to config/db.js
