+++ server.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const mongoose = require("mongoose");
 const cors = require("cors");
+const connectDB = require("./config/db");
 const app = express();
 require("dotenv").config();
 app.use(express.json());
@@ -8,24 +9,9 @@ app.use(cors());
 
 // Use environment variables
 const PORT = process.env.PORT || 5000;
-const MONGO_URI = process.env.DB_URL;
-const clientOptions = {
-  serverApi: { version: "1", strict: true, deprecationErrors: true },
-};
 
-async function run() {
-  try {
-    await mongoose.connect(MONGO_URI, clientOptions);
-    await mongoose.connection.db.admin().command({ ping: 1 });
-    console.log(
-      "Pinged your deployment. You successfully connected to MongoDB!"
-    );
-  } finally {
-    // Ensures that the client will close when you finish/error
-    await mongoose.disconnect();
-  }
-}
-run().catch(console.dir);
+// Connect to database
+connectDB();
 
 app.get("/", (req, res) => {
   res.send("Hello World");
